Cast cond_init's type to select correct variant of the method; see Bugzilla #4201

+++ ACE/ace/OS_NS_Thread.inl
@@ -1538,7 +1538,7 @@ ACE_OS::sema_init (ACE_sema_t *s,
   if (ACE_OS::mutex_init (&s->lock_, type, name,
                           (ACE_mutexattr_t *) arg) == 0
       && (attributes == 0 ?
-            ACE_OS::cond_init (&s->count_nonzero_, type, name, arg) :
+          ACE_OS::cond_init (&s->count_nonzero_, (short)type, name, arg) :
             ACE_OS::cond_init (&s->count_nonzero_, *attributes, name, arg)) == 0
       && ACE_OS::mutex_lock (&s->lock_) == 0)
     {
@@ -1759,7 +1759,7 @@ ACE_OS::sema_init (ACE_sema_t *s,
   if (ACE_OS::mutex_init (&s->lock_, type, name,
                           (ACE_mutexattr_t *) arg) == 0
       && (attributes == 0 ?
-            ACE_OS::cond_init (&s->count_nonzero_, type, name, arg) :
+          ACE_OS::cond_init (&s->count_nonzero_, (short)type, name, arg) :
             ACE_OS::cond_init (&s->count_nonzero_, *attributes, name, arg)) == 0
       && ACE_OS::mutex_lock (&s->lock_) == 0)
     {
